feat: impl Class Builder
I used a consuming builder pattern this time around. It reduces the
number of required clones, and it forces the developer to be more
explicit when they want to reuse a builder.

As I am filling the `ui` method with more `Character` instantiation,
I am starting to notice a code smell. It has become particularly
potent around the `class::Builder` with all of its `.and_then`s.

I have also noticed that the program does not crash gracefully when an
issue occurs within the `ui` method. As the `panic!` ends the program
when it is invoked, the program never has a chance to call
`restore_terminal` to clean up the mess it made...