chore: remove `main.rs` symlink
The symlink was causing some really interesting bugs in my editor.
To start with, it was reporting all errors twice - which really is not
a big deal. The main issue was with renaming over files. For some
reason, it would always mess up the name, duplicating a portion of it
at the end.